Redirect to URL with trailing slash when appropriate
Redirect from 'path/to/resource' to 'path/to/resource/' when

* 'path/to/resource' is not an existing file, and
* 'path/to/resource/' exists as a directory with index page.

The index page must not be served under 'path/to/resource'
without trailing slash, otherwise relative URLs originating
from the index page can not be resolved properly and break.

Use a 301 (Moved Permanently) response for redirection, but with
Cache-Control header to allow for future changes of the site.